+++ Assets/Scripts/Game/GameManager/LevelGeneratorController.cs
@@ -9,6 +9,7 @@ using UnityEngine;
 using UnityEngine.SceneManagement;
 using Overlord.LevelGenerator.EvolutionaryAlgorithm;
 using Overlord.LevelGenerator.Manager;
+using Overlord.Events;
 
 namespace Game.GameManager
 {
