+++ Assets/Scripts/Overlord/LevelsGenerator/LevelGeneratorManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9575ca784fb76414d8101084a8b89c7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
